% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.bins}
\alias{create.bins}
\title{create.bins}
\usage{
create.bins(binSize, genomeLength)
}
\arguments{
\item{binSize}{Numeric value of the desired size for each bin}

\item{genomeLength}{Numeric Length of the genome}
}
\value{
GRanges object
}
\description{
Create a GRanges object given the length of the genome and the desired size for each bin
}

