% Generated by roxygen2 (4.0.1): do not edit by hand
\name{density.reads.per.strand.ratio}
\alias{density.reads.per.strand.ratio}
\title{density.reads.per.strand.ratio}
\usage{
density.reads.per.strand.ratio(bins, reads)
}
\arguments{
\item{bins}{GRanges / GAlignments object}

\item{reads}{GRanges / GAlgnments, it need to contain the strand}
}
\value{
density object
}
\description{
Given a predefined bin partition of the genome, this counts the number of reads with forward and backward strand in each bin and obtaines the ratio for the forward strand defined as (counts_F + 1)/(counts_F + counts_R + 2)
}
\seealso{
\code{\link{create.bin}}
}

