% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resume.samples}
\alias{resume.samples}
\title{resume.samples}
\usage{
resume.samples(edsn = NULL, cult = NULL, ip = NULL, phase = NULL,
  growth = NULL, rif = NULL, rep = NULL, seq = NULL)
}
\arguments{
\item{edsn}{Identifier of experiment}

\item{cult}{Cultivation number of the experiment}

\item{ip}{Inmunoprecipitate used in the experiment}

\item{phase}{Phase used, could be Exponential or Stationary}

\item{Growth}{condition of the experiment, could be Aerobic or Anaerobic}

\item{rif}{condition of the experiment, could be 0 min or 20 min}

\item{rep}{Replicate the experiment, usually 1 or 2}

\item{seq}{Type of sequencing used, Exo, PET or SET}
}
\value{
Character with the logical structure used to filter the conditions table
}
\description{
Generates a string used to match ChIP-Seq PET and ChIP-Exo samples with same conditions
}

